+++ src/plugins/backup/index.ts
@@ -126,44 +126,40 @@ export class RxBackupState {
 
         await Promise.all(
             Object
-                .keys(this.database.collections)
-                .map(async (collectionName) => {
+                .entries(this.database.collections)
+                .map(async ([collectionName, collection]) => {
+                    const primaryKey = collection.schema.primaryPath;
                     const processedDocuments: Set<string> = new Set();
-                    const collection: RxCollection = this.database.collections[collectionName];
 
                     await this.database.requestIdlePromise();
 
                     if (!meta.collectionStates[collectionName]) {
-                        meta.collectionStates[collectionName] = {
-                            lastSequence: 0
-                        };
+                        meta.collectionStates[collectionName] = {};
                     }
-                    let lastSequence = meta.collectionStates[collectionName].lastSequence;
+                    let lastCheckpoint = meta.collectionStates[collectionName].checkpoint;
 
                     let hasMore = true;
                     while (hasMore && !this.isStopped) {
                         await this.database.requestIdlePromise();
-                        const changesResult = await collection.storageInstance.getChangedDocuments({
-                            sinceSequence: lastSequence,
-                            limit: this.options.batchSize,
-                            direction: 'after'
-                        });
-                        lastSequence = changesResult.lastSequence;
+                        const changesResult = await collection.storageInstance.getChangedDocumentsSince(
+                            this.options.batchSize ? this.options.batchSize : 0,
+                            lastCheckpoint
+                        );
+                        lastCheckpoint = changesResult.checkpoint;
+                        meta.collectionStates[collectionName].checkpoint = lastCheckpoint;
 
-                        meta.collectionStates[collectionName].lastSequence = lastSequence;
-
-                        const docIds: string[] = changesResult.changedDocuments
-                            .filter(changedDocument => {
+                        const docIds: string[] = changesResult.documents
+                            .map(docData => docData[primaryKey])
+                            .filter(id => {
                                 if (
-                                    processedDocuments.has(changedDocument.id)
+                                    processedDocuments.has(id)
                                 ) {
                                     return false;
                                 } else {
-                                    processedDocuments.add(changedDocument.id);
+                                    processedDocuments.add(id);
                                     return true;
                                 }
                             })
-                            .map(r => r.id)
                             .filter((elem, pos, arr) => arr.indexOf(elem) === pos); // unique
                         await this.database.requestIdlePromise();
 
@@ -199,10 +195,8 @@ export class RxBackupState {
                                     });
                                 })
                         );
-
                     }
-
-                    meta.collectionStates[collectionName].lastSequence = lastSequence;
+                    meta.collectionStates[collectionName].checkpoint = lastCheckpoint;
                     await setMeta(this.options, meta);
                 })
         );