+++ src/plugins/worker/non-worker.ts
@@ -123,10 +123,17 @@ export class RxStorageInstanceWorker<RxDocType> implements RxStorageInstance<RxD
             attachmentId
         );
     }
-    getChangedDocuments(options: ChangeStreamOnceOptions): Promise<{ changedDocuments: RxStorageChangedDocumentMeta[]; lastSequence: number; }> {
-        return this.internals.worker.getChangedDocuments(
+    async getChangedDocumentsSince(
+        limit: number,
+        checkpoint?: any
+    ): Promise<{
+        documents: RxDocumentData<RxDocType>[];
+        checkpoint?: any;
+    }> {
+        return this.internals.worker.getChangedDocumentsSince(
             this.internals.instanceId,
-            options
+            limit,
+            checkpoint
         );
     }
     changeStream(): Observable<EventBulk<RxStorageChangeEvent<RxDocumentData<RxDocType>>>> {