+++ src/types/rx-storage.interface.d.ts
@@ -259,21 +259,21 @@ export interface RxStorageInstance<
         attachmentId: string
     ): Promise<string>;
 
+
     /**
-     * Returns the ids of all documents that have been
-     * changed since the given sinceSequence.
+     * Returns the current (not the old!) data of all documents that have been changed AFTER the given checkpoint.
+     * This might return the same document multiple times when paginating over the checkpoint depending on the RxStorage.
+     * If the returned array does not reach the limit, it can be assumed that the "end" is reached.
+     * Also returns a new checkpoint which can be used to continue with the pagination.
+     * This is used by RxDB to known what has changed since X so these docs can be handled by the backup or the replication
+     * plugin.
      */
-    getChangedDocuments(
-        options: ChangeStreamOnceOptions
+    getChangedDocumentsSince(
+        limit: number,
+        checkpoint?: any
     ): Promise<{
-        changedDocuments: RxStorageChangedDocumentMeta[],
-        /**
-         * The last sequence number is returned in a separate field
-         * because the storage instance might have left out some events
-         * that it does not want to send out to the user.
-         * But still we need to know that they are there for a gapless pagination.
-         */
-        lastSequence: number;
+        documents: RxDocumentData<RxDocType>[];
+        checkpoint?: any;
     }>;
 
     /**